requirements.txt
pytest==7.4.0
pydantic==2.0.3
PyYAML==6.0
